+++ index.html
@@ -166,6 +166,6 @@
   </template>
 
 
-  <script type="text/javascript" src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
